refactor: LocationManager 관련 피드백 반영
- LocationManager 에서 좌표를 받아오는 기능 GeocoderManager 로 이동
- didUpdateLocations 에서 기존 completion 대신 델리게이트를 통해 좌표 전달하도록 변경. 관련 프로토콜 추가.
