+++ WeatherForecast/WeatherForecast/Managers/GeocoderManager.swift
@@ -5,7 +5,6 @@
 //  Created by YongHoon JJo on 2021/10/13.
 //
 
-import Foundation
 import CoreLocation
 
 enum GeocoderError: LocalizedError {
@@ -19,8 +18,11 @@ enum GeocoderError: LocalizedError {
     }
 }
 
-class GeocoderManager {
+class GeocoderManager: NSObject, CLLocationManagerDelegate {
     private let geocoder: GeocoderProtocol
+    private var debounceWorkItem: DispatchWorkItem?
+    
+    weak var delegate: GeocoderManagible?
     
     init(geocoder: GeocoderProtocol = CLGeocoder()) {
         self.geocoder = geocoder
@@ -44,6 +46,44 @@ class GeocoderManager {
             }
                                             
             completion(.success(address))
+        }
+    }
+}
+
+// MARK: - CLLocationManagerDelegate
+extension GeocoderManager {
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let currentLocation = locations.last else {
+            return
+        }
+        
+        debounceWorkItem?.cancel()
+        let requestWorkItem = DispatchWorkItem { [weak self] in
+            guard let this = self else { return }
+            this.delegate?.updateCoordiante(coordinate: currentLocation.coordinate)
+        }
+        
+        debounceWorkItem = requestWorkItem
+        DispatchQueue.global().asyncAfter(deadline: .now() + 0.3, execute: requestWorkItem)
+        
+        manager.stopUpdatingLocation()
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+        // TODOs: 좌표정보 직접 입력받거나 현재 위치 재설정 알럿 띄우기
+    }
+    
+    func locationManager(_ manager: CLLocationManager,
+                         didChangeAuthorization status: CLAuthorizationStatus) {
+        switch status {
+        case .authorizedWhenInUse:
+            manager.startUpdatingLocation()
+            
+        case .notDetermined:
+            manager.requestWhenInUseAuthorization()
+
+        @unknown default:
+            break
         }
     }
 }
